+++ built_ins/pwd.c
@@ -6,7 +6,7 @@
 /*   By: fshields <fshields@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 16:50:36 by fshields          #+#    #+#             */
-/*   Updated: 2024/02/22 16:56:11 by fshields         ###   ########.fr       */
+/*   Updated: 2024/02/28 08:52:27 by fshields         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int	ft_print_pwd(int fd)
 	char	*pwd;
 	
 	pwd = get_pwd();
-	if (write(fd, pwd, ft_strlen(pwd)) != ft_strlen(pwd))
+	if (write(fd, pwd, ft_strlen(pwd)) != (ssize_t) ft_strlen(pwd))
 		return (1);
 	if (write(fd, "\n", 1) != 1)
 		return (1);
